C15 (v0.3.3): Encrypting refactored
