+++ src/core/handlers/crypto/file-crypter-handler.js
@@ -1,113 +1,55 @@
 const { ipcMain } = require('electron');
 const fs = require('fs');
-const path = require('path');
-const crypto = require('crypto');
-const { randomUUID } = require('crypto');
+const {
+  buildFileMetadata,
+  encryptFile,
+  decryptFile,
+  buildEncodedFilePath,
+  buildDecodedFilePath,
+} = require('../../utils/cipher-utils')
 
-const RSA_BLOCK_SIZE = 512;
-const FILE_TYPE_ENDING = '==CRTF==';
-const FILE_TYPE_ENDING_SIZE = FILE_TYPE_ENDING.length;
-
-const IV_SIZE = 16;
-const SALT_SIZE = 32;
-const AES_ALGORITHM = 'aes-256-cbc';
-const KEY_LENGTH = 32;
-
-function generateRandomUUIDCryptoneFileName(dirPath) {
-  const uuidFileName = `${randomUUID()}.crtn`;
-  const fullPath = path.join(dirPath, uuidFileName);
-  if (!fs.existsSync(fullPath)) {
-    return uuidFileName;
-  }
-  return generateRandomUUIDCryptoneFileName(dirPath);
-}
-
-function encryptContentWithAES(fileBuffer, password, iv, salt) {
-  const key = crypto.scryptSync(password, salt, KEY_LENGTH);
-  const cipher = crypto.createCipheriv(AES_ALGORITHM, key, iv);
-
-  return Buffer.concat([
-    cipher.update(fileBuffer),
-    cipher.final()
-  ]);
-}
-
-function decryptContentWithAES(encryptedBuffer, password, iv, salt) {
-  const key = crypto.scryptSync(password, salt, KEY_LENGTH);
-  const decipher = crypto.createDecipheriv(AES_ALGORITHM, key, iv);
-
-  return Buffer.concat([
-    decipher.update(encryptedBuffer),
-    decipher.final()
-  ]);
-}
-
-function isCryptoneEncoded(filePath) {
-  if (!fs.existsSync(filePath)) return false;
-  const fileData = fs.readFileSync(filePath);
-  const endBytes = fileData.slice(fileData.length - FILE_TYPE_ENDING_SIZE);
-  return endBytes.toString('ascii') === FILE_TYPE_ENDING;
-}
+const { isCryptoneEncoded } = require('../../utils/file-utils')
 
 function initializeFileCrypterHandler(mainWindow) {
 
   ipcMain.handle('encrypt-file', (event, cfile, password, publicKeyPath) => {
-    if (!fs.existsSync(cfile.path)) return;
-    const fileBuffer = fs.readFileSync(cfile.path);
-
-    const iv = crypto.randomBytes(IV_SIZE);
-    const salt = crypto.randomBytes(SALT_SIZE);
-
-    const fileNameBuffer = Buffer.from(cfile.name, 'utf8');
-    const metadata = Buffer.concat([iv, salt, fileNameBuffer]);
+    if (!fs.existsSync(cfile.path)) return {
+      success: false,
+      message: 'File does not exist',
+    };
 
     const rsaPublicKey = fs.readFileSync(publicKeyPath);
-    const rsaEncryptedMetadata = crypto.publicEncrypt(rsaPublicKey, metadata);
+    const fileBuffer = fs.readFileSync(cfile.path);
+    const fileMetadata = buildFileMetadata(cfile);
 
-    const aesEncryptedContent = encryptContentWithAES(fileBuffer, password, iv, salt);
-
-    const endingBuffer = Buffer.from(FILE_TYPE_ENDING, 'ascii');
-    const finalBuffer = Buffer.concat([
-      rsaEncryptedMetadata,
-      aesEncryptedContent,
-      endingBuffer
-    ]);
-
-    const outputDir = path.dirname(cfile.path);
-    const outputFileName = generateRandomUUIDCryptoneFileName(outputDir);
-    const outputFilePath = path.join(outputDir, outputFileName);
+    const outputFilePath = buildEncodedFilePath(cfile);
+    const finalBuffer = encryptFile(rsaPublicKey, fileBuffer, password, fileMetadata);
 
     fs.writeFileSync(outputFilePath, finalBuffer);
+    return {
+      success: true,
+      message: 'File encrypted',
+    }
   });
 
   ipcMain.handle('decrypt-file', (event, cfile, password, privateKeyPath) => {
-    if (!fs.existsSync(cfile.path)) {
-      throw new Error('File not found: ' + cfile.path);
+    if (!fs.existsSync(cfile.path)) return {
+      success: false,
+      message: 'File does not exist',
     }
-    if (!isCryptoneEncoded(cfile.path)) {
-      throw new Error('Not a Cryptone encrypted file');
+    if (!isCryptoneEncoded(cfile.path)) return {
+      success: false,
+      message: 'File is not Cryptone Encrypted',
     }
+    const {originalFileNameBuffer, decryptedFileContent} = decryptFile(cfile, privateKeyPath, password);
 
-    let fileBuffer = fs.readFileSync(cfile.path);
-    fileBuffer = fileBuffer.slice(0, fileBuffer.length - FILE_TYPE_ENDING_SIZE);
-
-    const rsaEncryptedMetadata = fileBuffer.slice(0, RSA_BLOCK_SIZE);
-    const aesEncryptedContent = fileBuffer.slice(RSA_BLOCK_SIZE);
-
-    const rsaPrivateKey = fs.readFileSync(privateKeyPath);
-    const decryptedMetadata = crypto.privateDecrypt(rsaPrivateKey, rsaEncryptedMetadata);
-
-    const iv = decryptedMetadata.slice(0, IV_SIZE);
-    const salt = decryptedMetadata.slice(IV_SIZE, IV_SIZE + SALT_SIZE);
-    const originalFileNameBuffer = decryptedMetadata.slice(IV_SIZE + SALT_SIZE);
-
-    const outputDir = path.dirname(cfile.path);
-    const originalFileName = originalFileNameBuffer.toString('utf8');
-    const outputFilePath = path.join(outputDir, originalFileName);
-
-    const decryptedFileContent = decryptContentWithAES(aesEncryptedContent, password, iv, salt);
-
+    const outputFilePath = buildDecodedFilePath(cfile, originalFileNameBuffer);
     fs.writeFileSync(outputFilePath, decryptedFileContent);
+
+    return {
+      success: true,
+      message: 'File decrypted',
+    }
   });
 
   ipcMain.handle('is-encrypted-file', (event, cfile) => {
@@ -117,5 +59,4 @@ function initializeFileCrypterHandler(mainWindow) {
 
 module.exports = {
   initializeFileCrypterHandler,
-  isCryptoneEncoded,
 };
